Be stricter about identifying reference types.
TypeReferenceOwnership here is a bitmap indicating which of
the various ownership modifiers (weak, unowned, unmanaged)
were present on the field declaration.  Making this a bitmap
ensures that we can accurately represent the data that may
have been provided but we have to be careful when we interpret
the result.

This changes the `isWeak`, `isUnowned`, `isUnmanged` checks to
only return true if exactly one bit is set.  Combinations of bits
will not be accepted.

Also, add `isStrong` to check for the default ownership with
no modifiers.