Sema: Split off TypeCheckDeclPrimary.cpp from TypeCheckDecl.cpp
The new file contains the implementation of typeCheckDecl() and
various utility functions that it uses. This code runs for
declarations in primary files only.

TypeCheckDecl.cpp now mostly consists of evaluate() implementations
for requests together with some utility functions.

A few of the new function prototypes I added to TypeCheckDecl.h
are indicative of some code that could be refactored further.
Any utility functions shared by both TypeCheckDecl.cpp and
TypeCheckDeclPrimary.cpp should be wrapped up in requests in
order for the results to be cached.