Factor a transform visitor out of IsBindableVisitor.
For substituted function type lowering, we want to take every structural position where a
concrete type matches the original generic type in an interface, and replace it with an
independent generic parameter. The first bit, matching up structural positions to generic
parameters, is more or less what `Type::isBindableTo` already does, so we can adapt its
visitor to take a callback that optionally substitutes into the type being walked based
on the generic argument and binding in the given pair of types.