Add regression test for rdar://57265336
There was a logic error causing us to place nominal types on the
DeclsToFinalize list in the specific case where they were members
of an extension of a class.

Of course this kind of thing is indicative of bad design, and now,
the whole DeclsToFinalize list is gone. Let's add a regression
test for posterity.