[windows] Adapt some Dependencies tests to Windows.
It seems that redirecting the stderr and then the stdout in Lit might
not work like in Unix and the stderr might be redirected to the original
stdout instead. To avoid that, remove the usage of cat, and redirect the
stdout to a temporal file, and then the stderr there.

Also, Windows do not lock the access to the file descriptor for stdout,
so stdout and stderr redirected to the same place might overlap. To
avoid partial matches, redirect stderr of some failing test to a
temporal file, in order to match only over the output (which seems to be
the only thing trying to be match).

Finally, it seems that Windows libc, and Darwin/Linux libc do not match
in the capitalization of some error strings. Use a pattern to match both
lower case and upper case.