build: separate out host build of SwiftRemoteMirror
Separate out the host build of SwiftRemoteMirror from the "target" build
(the host for the standard library may be different from the host for
the compiler).  Restructure the build to ensure that we use the correct
compiler for building the SwiftRemoteMirror for the host.  This fixes a
build issue when building for Linux AArch64.