Redo ObjC Override Checking
The general problem with this approach is that the clang importer is part of the cache-fill for name lookup.  This means that qualified lookups it runs will be re-entrant and result in validation-order-dependent behaviors.

The changed test case demonstrates this effect.  The changed CHECK-lines are because the old re-entrant override checking used to bounce around performing nested lookups until one of them skipped the override check and imported the root member.  It just so happened that this means that categories in the first header passed to Clang won and got to install their members.  What we really want is the opposite: categories in subsequent headers should be able to stomp categories in earlier headers.  This re-enables the pattern whereby a private/bridging header containing category overrides that refine the behavior and types of properties in public headers.  It also breaks a common cycle blocking requestification of lookupDirect