Resolve CodingKeys before visiting member decls in the primary
Force this check to occur so that the interface types of members can successfully make reference to CodingKeys.  This is an egregious emergent behavior of Codable that currently forces two protocol conformance checks just to synthesize a type witness.  We can definitely get away with just creating the witness up front.

rdar://56844567