Move semantic member synthesis into the type checker's lookup utilities
Pull these calls out of lookup itself to prevent re-entrancy.  Synthesizing these members can force a conformance check that can ultimately wind up back in lookup.  Pull the forcing out of every single qualified lookup and just install these members in the TypeChecker's entrypoints to lookup.

The upside is we can localize this gross hack, the downside is, of course, that we have to smear it across all of the lookup entrypoints.