[windows] Switch ICU4C from SourceForge to GitHub.
Use GitHub download instead of SourceForge for ICU4C binary package, since SF fails from time to time and the build fails because of it.

Also remove the usage of -z, since Community CI deletes the previous file anyway, so no point in checking for a cached copy, when the only thing that would happen is an annoying message in the build logs.