Minimize Make-Style Dependencies
When -emit-dependencies is switched on, the Frontends are each tasked
with emitting Make-style dependencies files describing the full Swift
compilation graph and any external dependencies picked up by the Clang
Importer. Each output path was given a make target, and each target was
given a full copy of the dependencies of the swift compilation session.

This is a tremendous amount of duplication of a tremendous amount of
dependencies.  There can often be hundreds of dependencies for a single
target, and each output file's target would recieve its own copy. This
puts a significant strain on the build system, which is parsing a linear
number of these dependency files in incremental mode. (Given the build
system operates at the level of driver invocations, it's questionable
whether we even need to emit these files with content at all, but that's
a separate issue). When each output file adds all of its dependencies,
that means we have a quadratic increase in the parsing load.

Aim to compress these make-style dependency files by nominating
a primary output as a most-preferred target that recieves the entire
compilation graph.  The remaining targets, rather than copy the build
dependencies, simply depend on that target being rebuilt.

This allows us to cut down the size of make-style dependencies by
(1/# of outputs).  When multiplied by the number of files in a large
incremental codebase, as in SR-11927, the savings can be as much as
50% of incremental build time.

We can reduce this further by nominating a single frontend job to be the
Blessed One That Emits The Dependencies since Make requires we emit
a complete graph (as in K^n) so everybody sees the same internal
dependencies, and there's no way for a user to customize
individual frontend invocations as with Clang so everybody should see
the same external dependencies.