Don't re-force member loading in already-forced class hierarchies
When deserializing a the (named) members of a class hierarchy, we
needn't walk all the way up the hierarchy, just into our superclass.
This will, in turn, walk to its superclass, etc.  This is as opposed to
this work being repeated at each level of the recursive hierarchy.
Should be a small perf win over the last patch.